Fix for WFMP-197, Move PluginProgressTracker class to core module
