+++ plugin/src/main/java/org/wildfly/plugin/provision/AbstractProvisionServerMojo.java
@@ -55,7 +55,7 @@ import static org.wildfly.plugin.core.Constants.PLUGIN_PROVISIONING_FILE;
 import static org.wildfly.plugin.core.Constants.STANDALONE_XML;
 import org.wildfly.plugin.core.FeaturePack;
 import org.wildfly.plugin.core.MavenRepositoriesEnricher;
-
+import org.wildfly.plugin.core.PluginProgressTracker;
 
 /**
  * Provision a server
